// src/shared/models/test-category.ts
export enum TestCategory {
  A = 'A',
  A1 = 'A1',
  A2 = 'A2',
  ADI2 = 'ADI2',
  ADI3 = 'ADI3',
  AM = 'AM',
  B = 'B',
  B1 = 'B1',
  BE = 'B+E',
  C = 'C',
  C1 = 'C1',
  C1E = 'C1+E',
  CCPC = 'CCPC',
  CE = 'C+E',
  D = 'D',
  D1 = 'D1',
  D1E = 'D1+E',
  DCPC = 'DCPC',
  DE =  'D+E',
  EUA1M1 = 'EUA1M1',
  EUA1M2 = 'EUA1M2',
  EUA2M1 = 'EUA2M1',
  EUA2M2 = 'EUA2M2',
  EUAM1 = 'EUAM1',
  EUAM2 = 'EUAM2',
  EUAMM1 = 'EUAMM1',
  EUAMM2 = 'EUAMM2',
  F = 'F',
  G = 'G',
  H = 'H',
  K = 'K',
  SC = 'SC',
}